Move template c code from source code to template files.
